Add creation date to exercise notes in JSON output
Add creation date to exercise notes in JSON output

Updated the JSON output to include the creation date extracted from the HTML title of each note. The 'DateTime' field is now populated based on the date found in the title of the associated HTML file. If the date is successfully parsed, it will be included in the exercise data; otherwise, it will be null.